+++ Client.Test/InfluxDbClientFactoryTest.cs
@@ -2,11 +2,14 @@ using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.IO;
+using System.Net;
+using System.Net.Http;
 using System.Reflection;
 using System.Security.Cryptography.X509Certificates;
 using InfluxDB.Client.Api.Client;
 using InfluxDB.Client.Core;
 using InfluxDB.Client.Core.Exceptions;
+using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
 namespace InfluxDB.Client.Test
@@ -544,6 +547,34 @@ AJvDAFTSr5A9GSjJ3OyIeKoI8Q6xuaQBitpZR90P/Ah/Ymg490rpXavk";
 
             var apiClient = GetDeclaredField<ApiClient>(_client.GetType(), _client, "_apiClient");
             Assert.AreEqual(certificateCollection, apiClient.RestClientOptions.ClientCertificates);
+        }
+
+        [Test]
+        public void InjectHttpClient()
+        {
+            var options = new InfluxDBClientOptions("http://localhost:8086")
+            {
+                Token = "my-token"
+            };
+
+            var services = new ServiceCollection();
+
+            services.AddHttpClient();
+            services.AddTransient(p =>
+            {
+                var httpClientFactory = p.GetService<IHttpClientFactory>();
+                options.HttpClient = httpClientFactory.CreateClient();
+                return new InfluxDBClient(options);
+            });
+
+            var builder = services.BuildServiceProvider();
+
+            _client = builder.GetRequiredService<InfluxDBClient>();
+
+            var restClient = GetDeclaredField<ApiClient>(_client.GetType(), _client, "_apiClient").RestClient;
+
+            Assert.AreEqual(options.HttpClient,
+                GetDeclaredField<HttpClient>(restClient.GetType(), restClient, "<HttpClient>k__BackingField"));
         }
 
         private static T GetDeclaredField<T>(IReflect type, object instance, string fieldName)